+++ zokrates_bellman/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod groth16;
 pub mod plonk;
 pub mod solidity_renderer;
+pub mod plonk_proving_scheme;
 
 extern crate bellman_ce as bellman;
