+++ zokrates_proof_systems/src/scheme/mod.rs
@@ -7,13 +7,10 @@ use zokrates_field::Field;
 pub mod gm17;
 pub mod groth16;
 pub mod marlin;
-pub mod plonk;
 
 pub use self::gm17::GM17;
 pub use self::groth16::G16;
 pub use self::marlin::Marlin;
-pub use self::plonk::Plonk;
-
 pub trait Scheme<T: Field>: Serialize + fmt::Debug + Clone {
     const NAME: &'static str;
